frontend/requirements.txt
streamlit==1.36.0
streamlit_pdf_viewer==0.0.14
pandas
plotly==5.22.0
azure-storage-blob==12.20.0
azure-cosmos==4.7.0
python-dotenv==1.0.1
azure-identity==1.17.1